Make ICompactSerializer take a DataOutput instead of DataOutputStream as its second parameter.
This caused issues for the GMS messages which need to have access to DataOutputStream.size().
I couldn't figure out a good way around this, so I chose a bad one -- there's a new
ICompactStreamSerializer which takes the old style DataOutputStream.

The purpose for this commit is so that we can write from serializers directly into SequenceFiles
with no intermediate copy